refactored selector into its own file, added a way of creating a browser-monkey object if needed (it is for browser-monkey tests otherwise they seem to get confused). I think this is a bit more logical anyway
